squash: move internet dependent tests to internet
This commit moves only the parts of test-dgram-membership.js that
require an internet connection, leaving the others in place.
